LPS-52124 Rename
